+++ src/styles/StyledMenuSuffix.tsx
@@ -1,9 +1,10 @@
-import styled from '@emotion/styled';
+import styled, { CSSObject } from '@emotion/styled';
 
 interface StyledMenuSuffixProps {
   firstLevel?: boolean;
   collapsed?: boolean;
   transitionDuration?: number;
+  rootStyles?: CSSObject;
 }
 
 export const StyledMenuSuffix = styled.span<StyledMenuSuffixProps>`
@@ -11,4 +12,6 @@ export const StyledMenuSuffix = styled.span<StyledMenuSuffixProps>`
   margin-left: 5px;
   opacity: ${({ firstLevel, collapsed }) => (firstLevel && collapsed ? '0' : '1')};
   transition: opacity ${({ transitionDuration }) => transitionDuration}ms;
+
+  ${({ rootStyles }) => rootStyles};
 `;